+++ src/api/api.constants.ts
@@ -1,0 +1,15 @@
+/** The name attributes for form inputs. Set on the individual form inputs but also used on the server side to read the data from query params. */
+export enum QueryParamKey {
+  minFontSize = 'minFontSize',
+  minScreenWidth = 'minWidth',
+  minRatio = 'minRatio',
+  maxFontSize = 'maxFontSize',
+  maxScreenWidth = 'maxWidth',
+  maxRatio = 'maxRatio',
+  allSteps = 'steps',
+  baseStep = 'base',
+  namingConvention = 'prefix',
+  shouldUseRems = 'useRems',
+  roundingDecimalPlaces = 'decimals',
+  fontFamily = 'font',
+}
