+++ src/components/FluidTypeScaleCalculator/Form/GroupUseRems/GroupUseRems.tsx
@@ -1,4 +1,4 @@
-import { QueryParamKey } from '../../../../types';
+import { QueryParamKey } from '../../../../api/api.constants';
 import Checkbox from '../../../Checkbox/Checkbox';
 import { useFormState } from '../../FluidTypeScaleCalculator.context';
