+++ src/types.ts
@@ -78,22 +78,6 @@ export type FormAction =
       payload: FormState['fontFamily'];
     };
 
-/** The name attributes for form inputs. Set on the individual form inputs but also used on the server side to read the data from query params. */
-export enum QueryParamKey {
-  minFontSize = 'minFontSize',
-  minScreenWidth = 'minWidth',
-  minRatio = 'minRatio',
-  maxFontSize = 'maxFontSize',
-  maxScreenWidth = 'maxWidth',
-  maxRatio = 'maxRatio',
-  allSteps = 'steps',
-  baseStep = 'base',
-  namingConvention = 'prefix',
-  shouldUseRems = 'useRems',
-  roundingDecimalPlaces = 'decimals',
-  fontFamily = 'font',
-}
-
 export type WithDispatch = {
   /** A dispatch function to update the app state. */
   dispatch: Dispatch<FormAction>;
